+++ src/components/board/components/sortingContainer.js
@@ -1,0 +1,13 @@
+/**
+ * Создание разметки блока сортировки задач
+ * @return {string} блок сортировки задач
+ */
+export const createSorting = () => {
+  return (`
+    <div class="board__filter-list">
+      <a href="#" class="board__filter" data-sort-type="default">SORT BY DEFAULT</a>
+      <a href="#" class="board__filter" data-sort-type="date-up">SORT BY DATE up</a>
+      <a href="#" class="board__filter" data-sort-type="date-down">SORT BY DATE down</a>
+    </div>
+  `);
+};
