+++ src/components/button-load-more/index-button-load-more.js
@@ -1,3 +1,5 @@
+import {loadMoreClickHandler} from "../listeners";
+
 /**
  * Создание шаблона кнопки показа оставшихся задач
  * @return {string} - кнопка
